Refactor user-settings and move it to new module
