--- src/components/user-settings.tsx
@@ -1,481 +1,0 @@
-import { DropdownMenu, Icon, MenuGroup, MenuItem, Spinner, TabPanel } from '@wordpress/components';
-import { sprintf } from '@wordpress/i18n';
-import { moreVertical, trash } from '@wordpress/icons';
-import { useI18n } from '@wordpress/react-i18n';
-import { useCallback, useState, useEffect } from 'react';
-import { Snapshot } from 'cli/lib/appdata';
-import Button from 'src/components/button';
-import { EditorPicker } from 'src/components/editor-picker';
-import { Gravatar } from 'src/components/gravatar';
-import { LanguagePicker } from 'src/components/language-picker';
-import Modal from 'src/components/modal';
-import offlineIcon from 'src/components/offline-icon';
-import ProgressBar from 'src/components/progress-bar';
-import { Tooltip } from 'src/components/tooltip';
-import { WordPressLogo } from 'src/components/wordpress-logo';
-import { WPCOM_PROFILE_URL } from 'src/constants';
-import { useAuth } from 'src/hooks/use-auth';
-import { useEditorData } from 'src/hooks/use-editor-data';
-import { useFeatureFlags } from 'src/hooks/use-feature-flags';
-import { useI18nData } from 'src/hooks/use-i18n-data';
-import { useIpcListener } from 'src/hooks/use-ipc-listener';
-import { useOffline } from 'src/hooks/use-offline';
-import { usePromptUsage } from 'src/hooks/use-prompt-usage';
-import { useSnapshots } from 'src/hooks/use-snapshots';
-import { cx } from 'src/lib/cx';
-import { getIpcApi } from 'src/lib/get-ipc-api';
-
-const UserInfo = ( {
-	user,
-	onLogout,
-}: {
-	user?: { displayName: string; email: string };
-	onLogout: () => void;
-} ) => {
-	const { __ } = useI18n();
-	return (
-		<div className="flex w-full gap-5">
-			<div className="flex w-full items-center gap-3">
-				<Tooltip text={ __( 'Edit profile' ) } placement="bottom">
-					<Button
-						onClick={ () => getIpcApi().openURL( WPCOM_PROFILE_URL ) }
-						aria-label={ __( 'Edit profile' ) }
-						variant="icon"
-					>
-						<Gravatar detailedDefaultImage size={ 32 } isBlack />
-					</Button>
-				</Tooltip>
-				<div className="flex flex-col">
-					<span className="overflow-ellipsis">{ user?.displayName }</span>
-					<span className="text-a8c-gray-700 text-[10px] leading-[10px]">{ user?.email }</span>
-				</div>
-			</div>
-			<Button variant="secondary" className="!gap-3" onClick={ onLogout }>
-				{ __( 'Log out' ) }
-			</Button>
-		</div>
-	);
-};
-
-const SnapshotInfo = ( {
-	siteCount,
-	siteLimit,
-	isDisabled,
-	onRemoveSnapshots,
-	isDeleting = false,
-}: {
-	siteCount: number;
-	siteLimit: number;
-	isDisabled?: boolean;
-	onRemoveSnapshots: () => void;
-	isDeleting?: boolean;
-} ) => {
-	const { __ } = useI18n();
-	const { snapshotCreationBlocked } = useSnapshots();
-	const menuItemStyles = cx(
-		'[&_span]:min-w-0 [&_span]:p-[1px]',
-		isDisabled &&
-			'[&.components-button:disabled]:cursor-not-allowed [&.components-button]:aria-disabled:cursor-not-allowed'
-	);
-	const isOffline = useOffline();
-	const offlineMessage = __( 'Deleting preview sites requires an internet connection.' );
-	return (
-		<div className={ cx( 'flex flex-col', ! snapshotCreationBlocked && 'gap-3' ) }>
-			<h2 className="a8c-label-semibold">{ __( 'Preview sites' ) }</h2>
-			<div className="flex gap-3 flex-row items-center w-full">
-				{ snapshotCreationBlocked ? (
-					<div className="text-a8c-gray-70">
-						{ __( 'Preview sites are not available for your account.' ) }
-					</div>
-				) : (
-					<>
-						<div className="flex w-full flex-col gap-2">
-							<div className="flex w-full flex-row justify-between gap-8">
-								<div className="flex flex-row items-center text-right">
-									{ isDeleting && <Spinner className="!mt-0 !mx-2" /> }
-									<span className="text-a8c-gray-70">
-										{ sprintf( __( '%1s of %2s active preview sites' ), siteCount, siteLimit ) }
-									</span>
-								</div>
-							</div>
-							<ProgressBar value={ siteCount } maxValue={ siteLimit } />
-						</div>
-						<DropdownMenu
-							className={
-								'ml-auto flex items-center [&_button:first-child]:p-0 [&_button:first-child]:min-w-6 [&_button:first-child]:h-6'
-							}
-							popoverProps={ { position: 'bottom left', resize: true } }
-							icon={ <Icon icon={ moreVertical }></Icon> }
-							size={ 24 }
-							label={ __( 'More options' ) }
-						>
-							{ ( { onClose }: { onClose: () => void } ) => {
-								return (
-									<MenuGroup>
-										<Tooltip
-											disabled={ ! isOffline }
-											icon={ offlineIcon }
-											text={ offlineMessage }
-											placement="bottom"
-										>
-											<MenuItem
-												aria-description={ isOffline ? offlineMessage : '' }
-												/**
-												 * Because there is a single menu item, the `aria-disabled`
-												 * attribute is used rather than `disabled` so that screen
-												 * readers can focus the item to announce its disabled state.
-												 * Otherwise, dropdown toggle would toggle an empty menu.
-												 */
-												aria-disabled={ isDisabled }
-												icon={ trash }
-												iconPosition="left"
-												isDestructive
-												className={ menuItemStyles }
-												onClick={ () => {
-													if ( isDisabled ) {
-														return;
-													}
-
-													onRemoveSnapshots();
-													onClose();
-												} }
-											>
-												{ __( 'Delete all preview sites' ) }
-											</MenuItem>
-										</Tooltip>
-									</MenuGroup>
-								);
-							} }
-						</DropdownMenu>
-					</>
-				) }
-			</div>
-		</div>
-	);
-};
-
-function PromptInfo() {
-	const { __ } = useI18n();
-	const { promptCount, promptLimit } = usePromptUsage();
-	const isOffline = useOffline();
-
-	return (
-		<div className="flex gap-3 flex-col">
-			<h2 className="a8c-label-semibold">{ __( 'AI assistant' ) }</h2>
-			<div className="flex gap-3 flex-row items-center w-full">
-				<div className="flex w-full flex-col gap-2">
-					<div className="flex w-full flex-row justify-between gap-8 ">
-						<div className="flex flex-row items-center text-right">
-							<span className="text-a8c-gray-70">
-								{ isOffline
-									? __( "You're currently offline" )
-									: sprintf( __( '%1s of %2s monthly prompts used' ), promptCount, promptLimit ) }
-							</span>
-						</div>
-					</div>
-					<ProgressBar value={ promptCount } maxValue={ promptLimit } />
-				</div>
-				<div className="h-6 w-6"></div>
-			</div>
-		</div>
-	);
-}
-
-const NonAuthenticatedAccountTab = () => {
-	const { __ } = useI18n();
-	const { authenticate } = useAuth();
-	const isOffline = useOffline();
-	const offlineMessage = __( "You're currently offline." );
-
-	return (
-		<>
-			<div className="justify-between items-center w-full h-auto flex">
-				<WordPressLogo />
-				<Tooltip disabled={ ! isOffline } icon={ offlineIcon } text={ offlineMessage }>
-					<Button
-						aria-description={ isOffline ? offlineMessage : '' }
-						aria-disabled={ isOffline }
-						variant="primary"
-						onClick={ () => {
-							if ( isOffline ) {
-								return;
-							}
-							authenticate();
-						} }
-					>
-						{ __( 'Log in' ) }
-					</Button>
-				</Tooltip>
-			</div>
-		</>
-	);
-};
-
-const AccountTab = ( {
-	user,
-	logout,
-}: {
-	user?: { displayName: string; email: string };
-	logout: () => void;
-} ) => <UserInfo onLogout={ logout } user={ user } />;
-
-const PreferencesTab = ( { onClose }: { onClose: () => void } ) => {
-	const { __ } = useI18n();
-	const { locale: savedLocale, setLocale: setSavedLocale } = useI18nData();
-	const [ locale, setLocale ] = useState( savedLocale );
-
-	const { editor, handleEditorChange, saveEditorPreference, resetEditor, hasEditorChanges } =
-		useEditorData();
-
-	const savePreferences = async () => {
-		setSavedLocale( locale );
-		await saveEditorPreference();
-		onClose();
-	};
-
-	const cancelChanges = () => {
-		setLocale( savedLocale );
-		resetEditor();
-		onClose();
-	};
-
-	const hasChanges = locale !== savedLocale || hasEditorChanges;
-
-	return (
-		<>
-			<LanguagePicker value={ locale } onChange={ setLocale } />
-			<EditorPicker value={ editor } onChange={ handleEditorChange } />
-			<div className="mt-auto pt-6 flex justify-end gap-3">
-				<Button variant="tertiary" onClick={ cancelChanges }>
-					{ __( 'Cancel' ) }
-				</Button>
-				<Button variant="primary" onClick={ savePreferences } disabled={ ! hasChanges }>
-					{ __( 'Save' ) }
-				</Button>
-			</div>
-		</>
-	);
-};
-
-const UsageTab = ( {
-	loadingDeletingAllSnapshots,
-	activeSnapshotCount,
-	isLoadingAllSnapshots,
-	isLoadingSnapshotUsage,
-	allSnapshots,
-	isOffline,
-	snapshotQuota,
-	onRemoveSnapshots,
-}: {
-	loadingDeletingAllSnapshots: boolean;
-	activeSnapshotCount: number;
-	isLoadingAllSnapshots: boolean;
-	isLoadingSnapshotUsage: boolean;
-	allSnapshots: Pick< Snapshot, 'atomicSiteId' >[] | null;
-	isOffline: boolean;
-	snapshotQuota: number;
-	onRemoveSnapshots: () => void;
-} ) => (
-	<>
-		<SnapshotInfo
-			isDeleting={ loadingDeletingAllSnapshots }
-			isDisabled={
-				activeSnapshotCount === 0 ||
-				loadingDeletingAllSnapshots ||
-				isLoadingAllSnapshots ||
-				isLoadingSnapshotUsage ||
-				allSnapshots?.length === 0 ||
-				isOffline
-			}
-			siteCount={ activeSnapshotCount }
-			siteLimit={ snapshotQuota }
-			onRemoveSnapshots={ onRemoveSnapshots }
-		/>
-		<PromptInfo />
-	</>
-);
-
-export default function UserSettings() {
-	const { __ } = useI18n();
-	const [ deletedAllSnapshots, setDeletedAllSnapshots ] = useState( false );
-	const { isAuthenticated, authenticate, logout, user } = useAuth();
-	const { preferredEditor } = useFeatureFlags();
-	const { locale: savedLocale, setLocale: setSavedLocale } = useI18nData();
-	const {
-		allSnapshots,
-		activeSnapshotCount,
-		snapshotQuota,
-		isLoadingSnapshotUsage,
-		loadingDeletingAllSnapshots,
-		deleteAllSnapshots,
-		loadingServerSnapshots: isLoadingAllSnapshots,
-	} = useSnapshots();
-	const [ needsToOpenUserSettings, setNeedsToOpenUserSettings ] = useState( false );
-
-	const isOffline = useOffline();
-	const offlineMessage = __( "You're currently offline." );
-
-	const resetLocalState = useCallback( () => {
-		setNeedsToOpenUserSettings( false );
-	}, [] );
-
-	useIpcListener( 'user-settings', () => {
-		setNeedsToOpenUserSettings( ! needsToOpenUserSettings );
-	} );
-
-	useEffect( () => {
-		if ( deletedAllSnapshots && ! loadingDeletingAllSnapshots ) {
-			setDeletedAllSnapshots( false );
-			getIpcApi().showNotification( {
-				title: __( 'Delete Successful' ),
-				body: __( 'All preview sites have been deleted.' ),
-			} );
-		}
-	}, [ __, loadingDeletingAllSnapshots, deletedAllSnapshots ] );
-
-	const onRemoveSnapshots = useCallback( async () => {
-		if ( ! allSnapshots || allSnapshots.length === 0 ) {
-			return;
-		}
-
-		const CANCEL_BUTTON_INDEX = 0;
-		const DELETE_BUTTON_INDEX = 1;
-
-		const { response } = await getIpcApi().showMessageBox( {
-			type: 'warning',
-			message: __( 'Delete all preview sites' ),
-			detail: __(
-				'All preview sites that exist for your WordPress.com account, along with all posts, pages, comments, and media, will be lost.'
-			),
-			buttons: [ __( 'Cancel' ), __( 'Delete all' ) ],
-			cancelId: CANCEL_BUTTON_INDEX,
-		} );
-
-		if ( response === DELETE_BUTTON_INDEX ) {
-			await deleteAllSnapshots( allSnapshots );
-			setDeletedAllSnapshots( true );
-		}
-	}, [ allSnapshots, deleteAllSnapshots, __ ] );
-
-	if ( ! preferredEditor ) {
-		return (
-			<>
-				{ needsToOpenUserSettings && (
-					<Modal title={ __( 'Settings' ) } isDismissible onRequestClose={ resetLocalState }>
-						{ ! isAuthenticated && (
-							<div className="flex flex-col gap-6">
-								<div className="justify-between items-center w-full h-auto flex">
-									<WordPressLogo />
-									<Tooltip disabled={ ! isOffline } icon={ offlineIcon } text={ offlineMessage }>
-										<Button
-											aria-description={ isOffline ? offlineMessage : '' }
-											aria-disabled={ isOffline }
-											variant="primary"
-											onClick={ () => {
-												if ( isOffline ) {
-													return;
-												}
-												authenticate();
-											} }
-										>
-											{ __( 'Log in' ) }
-										</Button>
-									</Tooltip>
-								</div>
-								<div className="border-t border-[#F0F0F0] w-full"></div>
-								<LanguagePicker value={ savedLocale } onChange={ setSavedLocale } />
-							</div>
-						) }
-						{ isAuthenticated && (
-							<div className="gap-6 flex flex-col">
-								<UserInfo onLogout={ logout } user={ user } />
-								<div className="border-t border-[#F0F0F0] w-full"></div>
-								<div className="flex flex-col gap-6">
-									<LanguagePicker value={ savedLocale } onChange={ setSavedLocale } />
-									<SnapshotInfo
-										isDeleting={ loadingDeletingAllSnapshots }
-										isDisabled={
-											activeSnapshotCount === 0 ||
-											loadingDeletingAllSnapshots ||
-											isLoadingAllSnapshots ||
-											isLoadingSnapshotUsage ||
-											allSnapshots?.length === 0 ||
-											isOffline
-										}
-										siteCount={ activeSnapshotCount }
-										siteLimit={ snapshotQuota }
-										onRemoveSnapshots={ onRemoveSnapshots }
-									/>
-									<PromptInfo />
-								</div>
-							</div>
-						) }
-					</Modal>
-				) }
-			</>
-		);
-	}
-
-	const tabs = [
-		{
-			name: 'account',
-			title: __( 'Account' ),
-		},
-		{
-			name: 'preferences',
-			title: __( 'Preferences' ),
-		},
-		...( isAuthenticated
-			? [
-					{
-						name: 'usage',
-						title: __( 'Usage' ),
-					},
-			  ]
-			: [] ),
-	];
-
-	return (
-		<>
-			{ needsToOpenUserSettings && (
-				<Modal
-					title={ __( 'Settings' ) }
-					isDismissible
-					onRequestClose={ resetLocalState }
-					size="medium"
-					className={ cx(
-						'min-h-[350px]',
-						'[&_[role="document"]]:px-0',
-						'[&_[role="document"]]:mt-[64px]'
-					) }
-				>
-					<TabPanel className="w-full" tabs={ tabs } orientation="horizontal">
-						{ ( { name } ) => (
-							<div className="mt-6 px-8 flex flex-col gap-6">
-								{ name === 'account' &&
-									( isAuthenticated ? (
-										<AccountTab user={ user } logout={ logout } />
-									) : (
-										<NonAuthenticatedAccountTab />
-									) ) }
-								{ name === 'preferences' && <PreferencesTab onClose={ resetLocalState } /> }
-								{ name === 'usage' && isAuthenticated && (
-									<UsageTab
-										loadingDeletingAllSnapshots={ loadingDeletingAllSnapshots }
-										activeSnapshotCount={ activeSnapshotCount }
-										isLoadingAllSnapshots={ isLoadingAllSnapshots }
-										isLoadingSnapshotUsage={ isLoadingSnapshotUsage }
-										allSnapshots={ allSnapshots }
-										isOffline={ isOffline }
-										snapshotQuota={ snapshotQuota }
-										onRemoveSnapshots={ onRemoveSnapshots }
-									/>
-								) }
-							</div>
-						) }
-					</TabPanel>
-				</Modal>
-			) }
-		</>
-	);
-}
