+++ src/modules/user-settings/components/index.ts
@@ -1,0 +1,7 @@
+export { default as UserSettings } from './user-settings';
+export { AccountTab } from './account-tab';
+export { NonAuthenticatedAccountTab } from './non-authenticated-account-tab';
+export { PreferencesTab } from './preferences-tab';
+export { UsageTab } from './usage-tab';
+export { SnapshotInfo } from './snapshot-info';
+export { PromptInfo } from './prompt-info';
